requirements.txt
requests
wtforms
python-dotenv
Flask
mysql-connector-python
Flask-WTF
matplotlib
tqdm
Flask-Login